Wed, Oct 18, 2023, 8:26 PM -04:00

+++ chapter1/credit/credit2.c
@@ -5,7 +5,7 @@ int main(void)
 {
     long CCN;
     CCN = get_long("Enter Credit Card Number: ");
-    int ogCCN = CCN;
+    long ogCCN = CCN;
     // the ogCCN will be used later to help differentiate between 16 digits Mastercard and Visa
     int d = 0;
     int step1sum = 0;
@@ -49,6 +49,7 @@ int main(void)
         else if (d == 16)
         {
         // printf("VISA OR MASTERCARD\n");
+        // printf("%ld\n", ogCCN);
             if ((ogCCN / 1000000000000000) == 4)
             {
                 printf("VISA\n");
